Used /admin/projects instead of /api/v1/projects
