+++ src/routes/admin/jobs/+page.server.js
@@ -1,5 +1,5 @@
 import { listUsers } from '$lib/server/api/v1/auth_api.js';
-import { listProjects } from '$lib/server/api/v1/project_api';
+import { listProjects } from '$lib/server/api/admin_api.js';
 
 export async function load({ fetch }) {
 	/** @type {import('$lib/types').Project[]} */
